Added worker hiring

+++ src/main/java/com/alderaeney/farmcrashbackend/player/exceptions/NotEnoughMoneyException.java
@@ -1,10 +1,12 @@
 package com.alderaeney.farmcrashbackend.player.exceptions;
 
+import java.math.BigInteger;
+
 public class NotEnoughMoneyException extends RuntimeException {
     public Integer price;
-    public Integer availableMoney;
+    public BigInteger availableMoney;
 
-    public NotEnoughMoneyException(Integer price, Integer availableMoney) {
+    public NotEnoughMoneyException(Integer price, BigInteger availableMoney) {
         this.price = price;
         this.availableMoney = availableMoney;
     }
